+++ viewer/file_metadata.html
@@ -6,7 +6,7 @@
 </head>
 <body>
  <div id="photo_data" style="position: absolute; top: 0; left: 0px; padding-top: 0px;width: 220px; max-height: 100%; overflow: hidden; margin-left: 0; background-color: rgba(58,68,84,0.8); color:white; font-family: Calibri; line-height: 0.7;">
-<p style="margin-left: 5px;"><b>Numer drogi: </b>Mosaic51</p><p style="margin-left: 5px;"><b>Nazwa ulicy: </b>pano</p><p style="margin-left: 5px;"><b>Numer odcinka: </b>2022-10-26</p><p style="margin-left: 5px;"><b>Kilometraż: </b>09-04-08</p><p style="margin-left: 5px;"><b>Data: </b>2022-10-20 10:00:29</p></div>
+<p style="margin-left: 5px;"><b>Numer drogi: </b>Mosaic51</p><p style="margin-left: 5px;"><b>Nazwa ulicy: </b>pano</p><p style="margin-left: 5px;"><b>Numer odcinka: </b>2022-10-26</p><p style="margin-left: 5px;"><b>Kilometraż: </b>09-04-08</p><p style="margin-left: 5px;"><b>Data: </b>2022-10-20 10:00:39</p></div>
     </div>
 </body>
 </html>
